FIX: Remove trailing slash when storing URL
Why is this change required?

Before this change, a user can specify the URL as `http://example.com`
or `http://example.com/` but this will be treated as different URLs when
stored in the settings file. There is no need for us to make this
distinction and can potentially lead to a poor experience when using the
CLI since users are prompted to enter an API key again if they use
`http://example.com/` when the API key has already been stored under
`http://example.com`