Allow running system tests for multiple themes without having to rebuild
Before this commit, I was mounting the theme directory directing as a
volume mount to the docker container. However, volume mounts can only be
created when running a container so if the user wants to run tests for a
theme, we will have to stop the old container and start a new one and
setup the container all over again.

Instead of mounting the theme's directory directly into the container,
we instead just mount a tmp directory and copy the theme's spec folder
into the volumn mount when running the specs. This way, the volume mount
is not tied directly into a theme and the docker container can be used
to run the RSpec tests for any theme.