FEATURE: Support local discourse dev env for  `discourse_theme rspec`
What does this change do?

This commit changes the `discourse_theme rspec` command which allows the
user to configure a theme directory to use a local Discourse environment
to run the RSpec tests configured in the specified theme directory.

When the `discourse_theme rspec` command is first ran for a given
directory, the CLI will ask the user if they would like to configure a
local Discoure environment to run the tests. If the user selects 'Y',
the user will be prompted again to for the path to the local Discourse
repository. If the user selects 'n', the tests will run inside a Docker
container created using the `discourse/discourse_test:release` Docker
image. The user's preference for each theme directory is stored in the
settings file located at `~/.discourse_theme`. Similar to the other
commands, the user can reset this preference by running `discourse_theme
rspec <dir> --reset`.

Why is this change being introduced?

This change allows users that already has a local Discourse development
environment running to reuse the environment and thus not have the
overhead of running tests inside a Docker container. Also, running tests
inside a Docker container using the `discourse/discourse_test:release`
Docker image is very slow for OSX users on the ARM architecture as the
Docker image is not built for such users. This change also acts as a
bridge until we can ship an image for `discourse/discourse_test`
optimised for the ARM architecture.

Follow-up to c0c920280bef7869f0515f5e4220cf5cd3e408ef