FEATURE: Add `discourse_theme rspec <path to spec files>` command (#28)
* FEATURE: Add `discourse_theme rspec <path to spec files>` command

What does this change do?

This commit adds the `rspec` command to the `discourse_theme` CLI so
that themes can write system specs and run them easily without the need
to set up a local Discourse development environment.

Example:

```
discourse_theme rspec /path/to/theme
```

or

```
discourse_theme rspec /path/to/theme/spec/specific-folder
```

or

```
discourse_theme rspec /path/to/theme/spec/specific_rspec_file.rb
```

or

```
discourse_theme rspec
/path/to/theme/spec/specific_rspec_file_with_line_number.rb:3
```

Under the hood, the command runs a Docker container using the
`discourse/discourse_test:release` image and sets up the test
environment. This does mean the first run will incur some overhead and
there are ways to reduce this overhead but I will address them in
subsequent PRs. Once the Docker container is running, it simply runs
`bundle exec rspec` in the Docker container for the path that is given.

When running system specs, the `--headless` option can also be used to
run the system specs in headless mode. While we can't run a non-headless
chrome in the browser, we can run a chrome browser locally and ask the
Rails system test to connect to the localy browser. This is done by
having the CLI start a `chromedriver` instance locally and then passing
the `CAPYBARA_SERVER_HOST` and `CAPYBARA_REMOTE_DRIVER_URL` env
variables when running the RSpec command.