FIX: `discourse_theme rspec` command using Docker not copying theme
Why this change?

Currently when running the `discourse_theme rspec` command and opting to
run the tests via a Docker container does not copy the theme folder
correctly to the tmp folder. As a result, the tests will fail because
the theme folder is not accessible to the docker container running the
`rspec` command.