FEATURE: Add upload theme migrations prompt to `watch` commands (#38)
Why this change?

Prior to this change, running `discourse_theme watch` on a theme
directory will upload theme migration files. This means that the theme migration
is uploaded and ran on the server even if the author has not finished writing the migration.
To avoid this problem, `discourse_theme watch` will now prompt for the
user as to whether they would like to upload and run the pending
migrations in the theme. Example output is below:

```
» Using http://localhost:4200 from /home/tgxworld/.discourse_theme
» Using api key from /home/tgxworld/.discourse_theme
? How would you like to sync this theme? Sync with existing theme: 'DiscoTOC' (id:38)
» Uploading theme from /home/tgxworld/work/DiscoTOC
? Would you like to upload and run the following pending theme migration(s): /home/tgxworld/work/DiscoTOC/migrations/settings/0003-rename-settings.js

‣ Yes
  No
```