FEATURE: `watch` command prompts user if theme migrations should run (#40)
This commit is a partial revert of
7e2b90f5e13e8e30b3e1171fbf5183b61793c6f1 and
8e3b6c771130d5dee6297a32223fbf782ac68acc.

Previously, our strategy was to not upload theme migration files if the
user was not ready to run the theme migration files. However, this is a
problem because not uploading the theme migration files mean that we are
unable to write and run QUnit tests for the migrations.

Therefore, this commit changes our strategy to always upload theme
migrations file like before but adds the `skip_migrations` field to the
body of the upload theme request to indicate to the server that it
should not run theme migrations.